lint: add black formatter

+++ dongtai_web/views/messages_del.py
@@ -21,33 +21,33 @@ from django.utils.translation import gettext_lazy as _
 
 
 class _MessagesDelArgsSerializer(serializers.Serializer):
-    id = serializers.IntegerField(required=False,
-                                  default=None,
-                                  help_text=_('The id of Message'))
+    id = serializers.IntegerField(
+        required=False, default=None, help_text=_("The id of Message")
+    )
     all = serializers.NullBooleanField(
         required=False,
         default=False,
-        help_text=_('delete all messages when all is True'))
+        help_text=_("delete all messages when all is True"),
+    )
 
 
 class MessagesDelEndpoint(UserEndPoint):
     @extend_schema_with_envcheck(
         request=_MessagesDelArgsSerializer,
-        summary=_('Messages Delete'),
+        summary=_("Messages Delete"),
         description=_("Used by the user to delete the corresponding message"),
-        tags=[_('Messages')])
+        tags=[_("Messages")],
+    )
     def post(self, request):
         ser = _MessagesDelArgsSerializer(data=request.data)
         try:
             if ser.is_valid(True):
-                id_ = ser.validated_data['id']
-                all_ = ser.validated_data['all']
+                id_ = ser.validated_data["id"]
+                all_ = ser.validated_data["all"]
         except ValidationError as e:
             return R.failure(data=e.detail)
         if all_ is True:
-            IastMessage.objects.filter(
-                to_user_id=request.user.id).all().delete()
+            IastMessage.objects.filter(to_user_id=request.user.id).all().delete()
         else:
-            IastMessage.objects.filter(to_user_id=request.user.id,
-                                       pk=id_).delete()
-        return R.success(msg='success')
+            IastMessage.objects.filter(to_user_id=request.user.id, pk=id_).delete()
+        return R.success(msg="success")
